Backend: Change double to BigDecimal for StockItem.java

+++ backend/src/main/java/de/ffmjava/capstone/backend/stock/StockItem.java
@@ -1,10 +1,12 @@
 package de.ffmjava.capstone.backend.stock;
 
+import java.math.BigDecimal;
+
 public record StockItem(
         String id,
         String name,
         StockType type,
-        double amountInStock,
-        double pricePerKilo
+        BigDecimal amountInStock,
+        BigDecimal pricePerKilo
 ) {
 }
